requirements.txt
flask
python-dotenv
groq
gunicorn
google-auth